requirements.txt
certifi==2021.10.8
cloudinary==1.29.0
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.1.4
django-cloudinary-storage==0.3.0
django-cors-headers==3.6.0
django-environ==0.8.1
django-filter==21.1
djangorestframework==3.13.1
graphene==2.1.9
graphene-django==2.14.0
gunicorn==20.1.0
Markdown==3.3.6
Pillow==9.0.1
psycopg2-binary==2.9.3
djangorestframework==3.13.1